+++ docs/types/etls/devco-xls.md
@@ -1,6 +1,14 @@
 <!-- Generated by documentation.js. Update this documentation by updating the source code. -->
 
-## DevcoCsvTransform
+## getBudget
+
+Transform message (DEVCO XLS)
+
+### Parameters
+
+- `record`
+
+## DevcoXlsTransform
 
 Map fields for DEVCO producer, CSV file types
