--- services/ingestion/etl/devco/csv/test/stubs/record.json
@@ -1,47 +1,0 @@
-{
-  "ID": "2",
-  "Country": "Nigeria",
-  "Region": "West Africa",
-  "Lead Financier": "AFD",
-  "Funding Source": "AfiF",
-  "Project Title":
-    "Enhancing Vocational Training delivery for the power sector in Nigeria (NAPTIN)",
-  "CRIS No or ExCom Des": "C(2017) 3692",
-  "Project Type": "Technical Assistance and Capacity Building",
-  "Total Budget\n(Million Euro)": "50.7",
-  "Investment Grant (Million Euro)": "0",
-  "TA (Million Euro)": "8",
-  "Interest Rate Subsidy \n(Million Euro)": "",
-  "Guarantee (Million Euro)": "",
-  "Equity (Million Euro)": "",
-  "Budget Support (Million Euro)": "",
-  "Loan (Million Euro)": "",
-  "Total EU Contribution \n(Million Euro)": "8",
-  "Leverage": "6.34",
-  "1.1 Access on grid electricity ('000 people)": " n/a ",
-  "1.2 Access mini grid electricity ('000 people)": " n/a ",
-  "1.3 Access off-grid electricity ('000 people)": " n/a ",
-  "1.4 Inferred access (additional generation) ('000 people)": " n/a ",
-  "1.5 Inferred access (cross-border transmission) ('000 people)": " n/a ",
-  "1.6 Access to biomass/biogas clean cooking ('000 people)": " n/a ",
-  "1.7 Access to LPG/ethanol cooking ('000 people)": " n/a ",
-  "1.8 Electricity from renewables (GWh/year)": " n/a ",
-  "1.9 Renewable generation capacity (MW)": " n/a ",
-  "1.10 Electricity from energy efficiency (liberated capacity) (MW)": " n/a ",
-  "1.11 Transmission lines (km)": " n/a ",
-  "1.12 Distribution lines (km)": "6",
-  "1.13 Energy Savings (MWh/year)": " n/a ",
-  "1.14 GHG emissions avoided per year (ktons CO2eq)": " n/a ",
-  "1.15 No of direct jobs person/year (construction)": " n/a ",
-  "1.16 No of permanent jobs \n(operation)": " n/a ",
-  "Description and main objectives":
-    "The project aims to develop trainings curricula adapted to the market needs and  reinforce NAPTIN’s attractiveness by building new technical infrastructures and acquiring relevant technical equipment for the network of the 8 Regional Training centers and headquarters in Abuja.",
-  "Project Stage": "Feasibility",
-  "Contract Status": "In preparation",
-  "Starting Date": "",
-  "Ending Date": "",
-  "For AREI Projects (Endorsement) Y/N": "",
-  "GIS Localisation": "9.044737,\n7.487787",
-  "Comments":
-    "The project provides vocational training and therefore there is no energy generation involved."
-}
