+++ services/ingestion/etl/devco/xls/package.json
@@ -1,6 +1,6 @@
 {
   "private": true,
-  "name": "@eubfr/ingestion-etl-devco-csv",
+  "name": "@eubfr/ingestion-etl-devco-xls",
   "version": "0.5.0",
   "scripts": {
     "deploy": "sls deploy -v",
@@ -9,9 +9,8 @@
   "dependencies": {
     "@eubfr/lib": "^0.5.0",
     "@eubfr/logger-messenger": "^0.5.0",
-    "csv-parse": "4.3.1",
     "i18n-iso-countries": "3.7.8",
-    "stream-transform": "1.0.8"
+    "xlsx": "0.14.1"
   },
   "devDependencies": {
     "@babel/core": "7.2.2",
