+++ services/ingestion/etl/devco/xls/test/stubs/record.json
@@ -1,0 +1,45 @@
+{
+  "ID": 100,
+  "Country": "Chad",
+  "Region": "West Africa",
+  "Lead Financier": "EC",
+  "Funding Source": "EDF",
+  "Project Title": "Fourniture, installation, mise en SAV d'un an des équipements photovoltaïques au bénéfice de onze (11) délégations régionales du Ministère des Finances et du Budget et de trois (03) communes pilotes (Implementing Entity: Aplicaciones Tecnicas de la Energiasl)",
+  "CRIS No or ExCom Des": "2016 / 371004",
+  "Project Type": "Mini grid including generation, ",
+  "Total Budget\r\n(Million Euro)": 0.618036,
+  "Investment Grant (Million Euro)": 0.618036,
+  "TA (Million Euro)": 0,
+  "Interest Rate Subsidy \r\n(Million Euro)": 0,
+  "Guarantee (Million Euro)": 0,
+  "Equity (Million Euro)": 0,
+  "Budget Support (Million Euro)": 0,
+  "Loan (Million Euro)": 0,
+  "Total EU Contribution \r\n(Million Euro)": 0.618036,
+  "Leverage": 1,
+  "2.1 Direct and Inferred electricity access ('000 people)": 0,
+  "2.2 Clean cooking and fuel access ('000 people)": 0,
+  "2.3 Direct and Inferred access to energy ('000 people)": 0,
+  "2.4 Electricity from renewabes (GWh/year)": 0.24528000000000005,
+  "2.5 Reneable generation capacity (MW)": 0.14,
+  "2.6 Electricity generation capacity (MW)": 0.14,
+  "2.7 Transmission and distribution lines (km)": 0,
+  "2.8  GHG emissions avoided per year (ktons CO2eq)": 0.2,
+  "2.9 No of direct and permanent jons (construction and operation)": 0,
+  "BET1 \r\n(Access to energy)": 0,
+  "BET2 \r\n(Renewable energy generation and energy efficiency)": 0.14,
+  "BET3 \r\n(Contribution to the fight against climate change)": 0.2,
+  "EURF 1\r\n(No of people provided with access to electricity with EU support)": 0,
+  "EURF 2\r\n(Renewable energy production supported by the EU)": 0.24528000000000005,
+  "EURF 3\r\n(GHG emission avoided)": 0.2,
+  "SDG 7.1.1\r\nPercentage of population with access to electricity)": 0,
+  "SDG 7.1.2\r\n(Proportion of population with primary reliance on clean fuels and technology)": 0,
+  "SDG 7.2.1 \r\nRenewable energy share in the total final energy consumption)": 0.24528000000000005,
+  "SDG 7.3.1 \r\n(Energy intensity measured in terms of primary energy and GDP)": 0,
+  "SDG 8.3.1 \r\n(Proportion of informal employement in non-agriculture employment, by sex)": 0,
+  "Description and main objectives": "Fourniture, installation, mise en SAV d'un an des équipements photovoltaïques au bénéfice de onze (11) délégations régionales du Ministère des Finances et du Budget et de trois (03) communes pilotes",
+  "Project Stage": 0,
+  "Contract Status": "On going",
+  "Starting Date": 2018,
+  "Ending Date": 0
+}
